requirements.txt
openai==0.8.0
langdetect==1.0.9
python-dotenv==0.15.0
